RUnit {RUnit}	R Documentation

RUnit - Package Description

Description

This package models the common Unit Test framework for R and provides functionality to track results of test case execution and generate a summary report. It also provides tools for code inspection and thus for test case coverage analysis. The design is inspired by the popular JUnit unit test framework.

This package comes with a set of unit tests, serving as a test battery to check correct functioning against new R versions released as well as practical examples for writing test cases (see the ‘inst/unitTests’ subdirectory of the source package, or ‘unitTests’ contained in the binary package version).

Author(s)

Thomas Ko\"nig, Klaus Ju\"nemann & Matthias Burger

References

RUnit - A Unit Test Framework for R. useR! 2004 Vienna

See Also

See defineTestSuite, runTestSuite for unit testing or inspect and tracker for code inspection.

[Package RUnit version 0.4.32 Index]