options {RUnit}	R Documentation

RUnit options

Description

RUnit uses three options available via the global R options list

Details

RUnit specif options are added to R's global options list on package loading and removed again on pachage unloading.

Options used in RUnit

silent:

logical flag, default FALSE, sets the 'silent' argument for checkException. Allows to globally silence output from exception checks for all test suites excuted in one run.

verbose:

non-negative integer, default 1, 0: surpresses enclosing begin/end messages for each test case, 1: output enclosing begin/end messages for each test case

outfile:

NULL, connection or character, default NULL. If non-null has to be an open connection or a file name. Will be used to redirect all output to specified file/connection using sink. Connection is close after test suite execution call (via runTestSuite or runTestFile) has completed. If the file exists it is overwriten.

Author(s)

Matthias Burger

See Also

options, getOption, sink.

Examples

## Not run: 
##  quiet log output
ro <- getOption("RUnit")
ro$silent <- TRUE
ro$verbose <- 0L
options("RUnit"=ro)

## End(Not run)

[Package RUnit version 0.4.32 Index]