.setUp {RUnit}	R Documentation

Definition of RUnit Test Case code files.

Description

Either one or both functions have to be provided by the test case author, take precedence over the dummy definitions provided by the RUnit package and are called once for every test case identified.

Usage

.setUp()
.tearDown()

Details

To be written ...

Value

Functions do not return a value; called for their side effects.

Author(s)

Thomas König, Klaus Jünemann & Matthias Burger

See Also

runTestFile.

[Package RUnit version 0.4.32 Index]